Add files via upload
The Amazon Clone is a frontend-only web application developed using HTML, CSS, and JavaScript, mimicking the core design and layout of the popular e-commerce platform, Amazon. This project replicates the homepage, product listings, and shopping cart features, offering users a familiar interface for browsing products. While the app does not include backend functionality or responsive design, it demonstrates a strong understanding of frontend development principles, including structured layouts, custom styling, and interactive elements. The project provides a solid foundation that can be further extended with responsive features and backend integration in the future.